Update geospatial environment activation session
I added instructions to open the Jupyter Lab notebook and attempt importing a module (rioxarray). Rioxarray is used in Episode #1 and we often find that learners aren't able to import it. By adding this step to the Setup page, learners can attempt the import during office hours so we don't have to dedicate helpers to assisting with minor setup issues. Lines 274-281.